[refactor] 당첨 번호 발급 책임 변경
