+++ src/main/java/lotto/NumberPool.java
@@ -51,7 +51,22 @@ public class NumberPool {
     return LottoTickets.of(tickets);
   }
 
-  public static LottoNumber getNumber(int number) {
-    return integerLottoNumberMapper.get(number);
+  public static WinningNumber publishWinningNumber(String input) {
+    String[] tokenizedInputs = input.split(", ");
+    List<LottoNumber> numbers = new ArrayList<>();
+
+    for (String tokenizedInput : tokenizedInputs) {
+      int castedInput = Integer.parseInt(tokenizedInput);
+      validateValue(castedInput);
+      numbers.add(integerLottoNumberMapper.get(castedInput));
+    }
+
+    return WinningNumber.of(numbers);
+  }
+
+  private static void validateValue(int number) {
+    if (!(MINIMUM_LOTTO_NUMBER <= number && number <= MAXIMUM_LOTTO_NUMBER)) {
+      throw new IllegalArgumentException();
+    }
   }
 }
